beautifying, by stacktic

+++ external/dewey.c
@@ -226,8 +226,8 @@ static int
 vtest(arr_t *lhs, int tst, arr_t *rhs)
 {
 	int cmp;
-	int     c;
-	int     i;
+	unsigned int     c;
+	unsigned int     i;
 
 	for (i = 0, c = MAX(lhs->c, rhs->c) ; i < c ; i++) {
 		if ((cmp = DIGIT(lhs->v, lhs->c, i) - DIGIT(rhs->v, rhs->c, i)) != 0) {
@@ -307,7 +307,7 @@ dewey_match(const char *pattern, const char *pkg)
 	if (sep2) {
 		char ver[PKG_PATTERN_MAX];
 
-		strlcpy(ver, sep, MIN(sizeof(ver), sep2-sep+1));
+		strlcpy(ver, sep, MIN(sizeof(ver), (size_t)(sep2-sep+1)));
 		if (dewey_cmp(version, op, ver))
 			return 1;
 	}
